requirements.txt
aiohttp
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pdfplumber
python-docx
chainlit
